+++ CHANGELOG.md
@@ -36,6 +36,7 @@
 * Add paths to Settings
 * Highlight changes in the hex view
 * Ram Search: highlight results that would be removed by future search
+* Add `strace` to the list of debugger programs
 
 ### Changed